refactor: :construction: initialize the refactoring from theme to plugin
For the sake of simplicity, Brumes is being refactored from being an Obsidian theme to being an Obsidian plugin. Brumes will now provide presets for "Border",  a highly customizable theme, and add multiple tools to ease the process of creating a CoM case inside obsidian. This new version is more "utility" focused and more functionnal.