refactor: remove build folder and make the plugin build in the root folder
The esbuild.config.mjs was supposed to build the plugin in a potential demo vault that was supposed to be included directly in the plugin. This is no longer required since it's a better approach to follow what other plugins are doing: building at the root of the repository so it can just be cloned in the .obsidian/plugins directory.